Restructure Project, Set up Boilerplate, Basic readme
