+++ README.md
@@ -10,19 +10,20 @@ NDGI Project Geoharvester
 
 ###### Requirements:
 
-- Have node and npm installed (https://docs.npmjs.com/downloading-and-installing-node-js-and-npm)
 - Your favorite terminal
+- Have node and npm installed (https://docs.npmjs.com/downloading-and-installing-node-js-and-npm)
 
 ###### Run:
 
 1. cd into frontend folder ("geoharvester_frontend")
 2. run `npm i` to install dependencies (from package.json)
-3. run `npm start` to start the fronted on localhost (npm start is defined in package.json)
+3. run `npm start` to start the fronted on localhost (`npm start` is defined in package.json)
 
 #### Backend:
 
 ###### Requirements:
 
+- Your favorite terminal
 - Have pip installed
 - Have virtual environment in backend folder up and running (`python -m venv env &&  source ./env/bin/activate`)
 
@@ -30,8 +31,8 @@ NDGI Project Geoharvester
 
 1. cd into backend folder ("geoharvester_backend")
 2. run `pip install -r requirements.txt` to install dependencies
-3. cd into src (or where main.py is located)
-4. run `uvicorn main:app --reload` to start server on localhost
+3. run `uvicorn main:app --reload` to start server on localhost
+4. Check `localhost:8000`in your browser to verify that backend is running
 
 #### Troubleshooting:
